+++ bindings/python/src/pre_tokenizers.rs
@@ -6,6 +6,7 @@ use pyo3::types::*;
 use serde::ser::SerializeStruct;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
+use tk::normalizer::SplitDelimiterBehavior;
 use tk::pre_tokenizers::bert::BertPreTokenizer;
 use tk::pre_tokenizers::byte_level::ByteLevel;
 use tk::pre_tokenizers::delimiter::CharDelimiterSplit;
@@ -101,10 +102,10 @@ impl PreTokenizer for PyPreTokenizer {
 #[pymethods]
 impl PyPreTokenizer {
     #[staticmethod]
-    fn custom(pretok: PyObject) -> PyResult<Self> {
-        Ok(PyPreTokenizer {
+    fn custom(pretok: PyObject) -> Self {
+        PyPreTokenizer {
             pretok: PyPreTokenizerWrapper::Custom(CustomPreTokenizer::new(pretok)).into(),
-        })
+        }
     }
 
     fn __getstate__(&self, py: Python) -> PyResult<PyObject> {
@@ -243,14 +244,14 @@ impl PyByteLevel {
     }
 
     #[new]
-    #[args(add_prefix_space = "true")]
-    fn new(add_prefix_space: bool) -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((
+    #[args(add_prefix_space = "true", _kwargs = "**")]
+    fn new(add_prefix_space: bool, _kwargs: Option<&PyDict>) -> (Self, PyPreTokenizer) {
+        (
             PyByteLevel {},
             ByteLevel::default()
                 .add_prefix_space(add_prefix_space)
                 .into(),
-        ))
+        )
     }
 
     /// Returns the alphabet used by this PreTokenizer.
@@ -278,8 +279,8 @@ pub struct PyWhitespace {}
 #[pymethods]
 impl PyWhitespace {
     #[new]
-    fn new() -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((PyWhitespace {}, Whitespace::default().into()))
+    fn new() -> (Self, PyPreTokenizer) {
+        (PyWhitespace {}, Whitespace::default().into())
     }
 }
 
@@ -290,8 +291,8 @@ pub struct PyWhitespaceSplit {}
 #[pymethods]
 impl PyWhitespaceSplit {
     #[new]
-    fn new() -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((PyWhitespaceSplit {}, WhitespaceSplit.into()))
+    fn new() -> (Self, PyPreTokenizer) {
+        (PyWhitespaceSplit {}, WhitespaceSplit.into())
     }
 }
 
@@ -332,8 +333,8 @@ impl PySplit {
         ))
     }
 
-    fn __getnewargs__<'p>(&self, py: Python<'p>) -> PyResult<&'p PyTuple> {
-        Ok(PyTuple::new(py, &[" ", "removed"]))
+    fn __getnewargs__<'p>(&self, py: Python<'p>) -> &'p PyTuple {
+        PyTuple::new(py, &[" ", "removed"])
     }
 }
 
@@ -364,8 +365,8 @@ impl PyCharDelimiterSplit {
         ))
     }
 
-    fn __getnewargs__<'p>(&self, py: Python<'p>) -> PyResult<&'p PyTuple> {
-        Ok(PyTuple::new(py, &[" "]))
+    fn __getnewargs__<'p>(&self, py: Python<'p>) -> &'p PyTuple {
+        PyTuple::new(py, &[" "])
     }
 }
 
@@ -379,20 +380,27 @@ pub struct PyBertPreTokenizer {}
 #[pymethods]
 impl PyBertPreTokenizer {
     #[new]
-    fn new() -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((PyBertPreTokenizer {}, BertPreTokenizer.into()))
+    fn new() -> (Self, PyPreTokenizer) {
+        (PyBertPreTokenizer {}, BertPreTokenizer.into())
     }
 }
 
-/// This pre-tokenizer simply splits on punctuation as individual characters.`
+/// This pre-tokenizer simply splits on punctuation as individual characters.
+///
+/// Args:
+///     behavior (:class:`~tokenizers.SplitDelimiterBehavior`):
+///         The behavior to use when splitting.
+///         Choices: "removed", "isolated" (default), "merged_with_previous", "merged_with_next",
+///         "contiguous"
 #[pyclass(extends=PyPreTokenizer, module = "tokenizers.pre_tokenizers", name=Punctuation)]
-#[text_signature = "(self)"]
+#[text_signature = "(self, behavior=\"isolated\")"]
 pub struct PyPunctuation {}
 #[pymethods]
 impl PyPunctuation {
     #[new]
-    fn new() -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((PyPunctuation {}, Punctuation.into()))
+    #[args(behavior = "PySplitDelimiterBehavior(SplitDelimiterBehavior::Isolated)")]
+    fn new(behavior: PySplitDelimiterBehavior) -> (Self, PyPreTokenizer) {
+        (PyPunctuation {}, Punctuation::new(behavior.into()).into())
     }
 }
 
@@ -420,8 +428,8 @@ impl PySequence {
         ))
     }
 
-    fn __getnewargs__<'p>(&self, py: Python<'p>) -> PyResult<&'p PyTuple> {
-        Ok(PyTuple::new(py, &[PyList::empty(py)]))
+    fn __getnewargs__<'p>(&self, py: Python<'p>) -> &'p PyTuple {
+        PyTuple::new(py, &[PyList::empty(py)])
     }
 }
 
@@ -464,12 +472,16 @@ impl PyMetaspace {
     }
 
     #[new]
-    #[args(replacement = "PyChar('▁')", add_prefix_space = "true")]
-    fn new(replacement: PyChar, add_prefix_space: bool) -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((
+    #[args(replacement = "PyChar('▁')", add_prefix_space = "true", _kwargs = "**")]
+    fn new(
+        replacement: PyChar,
+        add_prefix_space: bool,
+        _kwargs: Option<&PyDict>,
+    ) -> (Self, PyPreTokenizer) {
+        (
             PyMetaspace {},
             Metaspace::new(replacement.0, add_prefix_space).into(),
-        ))
+        )
     }
 }
 
@@ -501,8 +513,8 @@ impl PyDigits {
 
     #[new]
     #[args(individual_digits = false)]
-    fn new(individual_digits: bool) -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((PyDigits {}, Digits::new(individual_digits).into()))
+    fn new(individual_digits: bool) -> (Self, PyPreTokenizer) {
+        (PyDigits {}, Digits::new(individual_digits).into())
     }
 }
 
@@ -516,8 +528,8 @@ pub struct PyUnicodeScripts {}
 #[pymethods]
 impl PyUnicodeScripts {
     #[new]
-    fn new() -> PyResult<(Self, PyPreTokenizer)> {
-        Ok((PyUnicodeScripts {}, UnicodeScripts::new().into()))
+    fn new() -> (Self, PyPreTokenizer) {
+        (PyUnicodeScripts {}, UnicodeScripts::new().into())
     }
 }