+++ bindings/python/src/utils/normalization.rs
@@ -92,7 +92,7 @@ impl PyRange<'_> {
 }
 
 #[derive(Clone)]
-pub struct PySplitDelimiterBehavior(SplitDelimiterBehavior);
+pub struct PySplitDelimiterBehavior(pub SplitDelimiterBehavior);
 
 impl FromPyObject<'_> for PySplitDelimiterBehavior {
     fn extract(obj: &PyAny) -> PyResult<Self> {
@@ -411,6 +411,20 @@ impl PyNormalizedStringRefMut {
     pub fn destroyed_error() -> PyErr {
         exceptions::PyException::new_err("Cannot use a NormalizedStringRefMut outside `normalize`")
     }
+
+    /// Provides a way to access a reference to the underlying NormalizedString
+    pub fn map_as_ref<F: FnOnce(&NormalizedString) -> U, U>(&self, f: F) -> PyResult<U> {
+        self.inner
+            .map(f)
+            .ok_or_else(PyNormalizedStringRefMut::destroyed_error)
+    }
+
+    /// Provides a way to access a mutable reference to the underlying NormalizedString
+    pub fn map_as_mut<F: FnOnce(&mut NormalizedString) -> U, U>(&mut self, f: F) -> PyResult<U> {
+        self.inner
+            .map_mut(f)
+            .ok_or_else(PyNormalizedStringRefMut::destroyed_error)
+    }
 }
 
 #[pymethods]
@@ -539,7 +553,7 @@ impl PyNormalizedStringRefMut {
 
     fn slice(&self, range: PyRange) -> PyResult<Option<PyNormalizedString>> {
         self.inner
-            .map(|n| slice(&n, &range))
+            .map(|n| slice(n, &range))
             .ok_or_else(PyNormalizedStringRefMut::destroyed_error)?
     }
 
@@ -552,7 +566,7 @@ impl PyNormalizedStringRefMut {
 
     fn for_each(&self, func: &PyAny) -> PyResult<()> {
         self.inner
-            .map(|n| for_each(&n, func))
+            .map(|n| for_each(n, func))
             .ok_or_else(PyNormalizedStringRefMut::destroyed_error)??;
         Ok(())
     }